refactor: move user::project types from lib to wire crate
