+++ lib/src/user/project.rs
@@ -1,83 +1,14 @@
 use crate::common::{request, request_bare, SerializableNone};
 use crate::error::ApiError;
-use crate::resources::FlavorGroupMinimal;
-use crate::user::UserMinimal;
 use anyhow::Context;
+use lrzcc_wire::user::{
+    Project, ProjectCreateData, ProjectCreated, ProjectDetailed,
+    ProjectModifyData,
+};
 use reqwest::blocking::Client;
 use reqwest::Url;
 use reqwest::{Method, StatusCode};
-use serde::{Deserialize, Serialize};
-use std::fmt::Display;
 use std::rc::Rc;
-use tabled::Tabled;
-
-#[derive(Clone, Debug, Deserialize, Serialize, Tabled)]
-pub struct Project {
-    pub id: u32,
-    pub name: String,
-    pub openstack_id: String, // UUIDv4 without dashes
-    pub user_class: u32,
-}
-
-impl Display for Project {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.write_str(&format!("Project(id={}, name={}", self.id, self.name))
-    }
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize, Tabled)]
-pub struct ProjectMinimal {
-    pub id: u32,
-    pub name: String,
-    pub user_class: u32,
-}
-
-impl Display for ProjectMinimal {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.write_str(&format!("Project(id={}, name={})", self.id, self.name))
-    }
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize, Tabled)]
-pub struct ProjectDetailed {
-    pub id: u32,
-    pub name: String,
-    pub openstack_id: String, // UUIDv4 without dashes
-    pub user_class: u32,
-    // TODO rethink list output in detailed structs:
-    // maybe we could have only the first few entries followed by ...
-    // in the output
-    #[tabled(skip)]
-    pub users: Vec<UserMinimal>,
-    #[tabled(skip)]
-    pub flavor_groups: Vec<FlavorGroupMinimal>,
-}
-
-impl Display for ProjectDetailed {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.write_str(&format!("Project(id={}, name={}", self.id, self.name))
-    }
-}
-
-// TODO can we merge this with ProjectDetailed via some enum
-// in the project field
-#[derive(Clone, Debug, Deserialize, Serialize, Tabled)]
-pub struct ProjectCreated {
-    pub id: u32,
-    pub name: String,
-    pub openstack_id: String, // UUIDv4 without dashes
-    pub user_class: u32,
-    #[tabled(skip)]
-    pub users: Vec<u32>,
-    #[tabled(skip)]
-    pub flavor_groups: Vec<u32>,
-}
-
-impl Display for ProjectCreated {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.write_str(&format!("Project(id={}, name={}", self.id, self.name))
-    }
-}
 
 pub struct ProjectApi {
     pub url: String,
@@ -137,24 +68,6 @@ impl ProjectListRequest {
     }
 }
 
-#[derive(Clone, Debug, Serialize)]
-struct ProjectCreateData {
-    name: String,
-    openstack_id: String, // UUIDv4
-    // #[serde(skip_serializing_if = "Option::is_none")]
-    user_class: Option<u32>,
-}
-
-impl ProjectCreateData {
-    fn new(name: String, openstack_id: String) -> Self {
-        Self {
-            name,
-            openstack_id,
-            user_class: None,
-        }
-    }
-}
-
 pub struct ProjectCreateRequest {
     url: String,
     client: Rc<Client>,
@@ -189,29 +102,6 @@ impl ProjectCreateRequest {
             Some(&self.data),
             StatusCode::CREATED,
         )
-    }
-}
-
-#[derive(Clone, Debug, Serialize)]
-struct ProjectModifyData {
-    id: u32,
-
-    #[serde(skip_serializing_if = "Option::is_none")]
-    name: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    openstack_id: Option<String>, // UUIDv4
-    #[serde(skip_serializing_if = "Option::is_none")]
-    user_class: Option<u32>,
-}
-
-impl ProjectModifyData {
-    fn new(id: u32) -> Self {
-        Self {
-            id,
-            name: None,
-            openstack_id: None,
-            user_class: None,
-        }
     }
 }
