+++ wire/src/user/mod.rs
@@ -2,5 +2,5 @@ mod project;
 #[allow(clippy::module_inception)]
 mod user;
 
-pub use project::ProjectApi;
-pub use user::UserApi;
+pub use project::*;
+pub use user::*;
